requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
attrs==21.4.0
blinker==1.4
Bottleneck==1.3.4
brotlipy==0.7.0
cachetools==5.0.0
certifi==2022.5.18.1
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
cryptography==37.0.1
frozenlist==1.3.0
google-auth==2.6.6
google-auth-oauthlib==0.5.1
gspread==5.3.0
httplib2==0.20.4
idna==3.3
multidict==6.0.2
numexpr==2.8.1
numpy==1.21.5
oauth2client==4.1.3
oauthlib==3.2.0
packaging==21.3
pandas==1.4.2
patsy==0.5.2
pip==21.2.4
pyasn1==0.4.8
pyasn1-modules==0.2.7
pycparser==2.21
PyJWT==2.4.0
pyOpenSSL==22.0.0
pyparsing==3.0.4
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2021.3
pyu2f==0.1.5
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
scipy==1.7.3
setuptools==62.3.2
six==1.16.0
statsmodels==0.13.2
typing_extensions==4.2.0
urllib3==1.26.9
wheel==0.37.1
yarl==1.7.2
